Update Storybook and tests
