+++ src/components/forms/AddressField/AddressField.stories.tsx
@@ -13,6 +13,7 @@ export default meta;
 export const Example = (args: AddressFieldProps) => <AddressField {...args} />;
 
 Example.args = {
+  type: "residence",
   children: (
     <div className="border border-gray-dim p-3 rounded-lg text-gray-dim">
       Children
