+++ src/components/forms/CheckboxGroupField/CheckboxGroupField.tsx
@@ -4,6 +4,9 @@ import {
   type CheckboxGroupProps,
   type CheckboxProps,
 } from "@/components/common";
+import type { UserFormDataField } from "@convex/constants";
+import { useFormContext } from "react-hook-form";
+import { Controller } from "react-hook-form";
 
 interface CheckboxOption extends CheckboxProps {
   label: string;
@@ -12,7 +15,7 @@ interface CheckboxOption extends CheckboxProps {
 
 export interface CheckboxGroupFieldProps extends CheckboxGroupProps {
   children?: React.ReactNode;
-  name: string;
+  name: UserFormDataField;
   label: string;
   labelHidden?: boolean;
   options: CheckboxOption[];
@@ -24,28 +27,41 @@ export function CheckboxGroupField({
   labelHidden,
   options,
   children,
+  ...props
 }: CheckboxGroupFieldProps) {
+  const { control } = useFormContext();
+
   return (
     <div className="@container flex flex-col gap-4">
-      <CheckboxGroup
-        label={!labelHidden ? label : undefined}
-        aria-label={label}
+      <Controller
+        control={control}
         name={name}
-        size="large"
-      >
-        <span className="italic text-gray-dim text-sm">
-          Select all that apply:
-        </span>
-        {options?.map(({ label, ...option }) => (
-          <Checkbox
-            key={option.value}
-            {...option}
+        defaultValue={[]}
+        render={({ field, fieldState: { error, invalid } }) => (
+          <CheckboxGroup
+            {...field}
+            label={!labelHidden ? label : undefined}
+            aria-label={label}
             size="large"
-            label={label}
-            card
-          />
-        ))}
-      </CheckboxGroup>
+            isInvalid={invalid}
+            errorMessage={error?.message}
+            {...props}
+          >
+            <span className="italic text-gray-dim text-sm">
+              Select all that apply:
+            </span>
+            {options?.map(({ label, ...option }) => (
+              <Checkbox
+                key={option.value}
+                {...option}
+                size="large"
+                label={label}
+                card
+              />
+            ))}
+          </CheckboxGroup>
+        )}
+      />
       {children}
     </div>
   );
