+++ src/components/forms/EmailField/EmailField.stories.tsx
@@ -13,6 +13,7 @@ export default meta;
 export const Example = (args: EmailFieldProps) => <EmailField {...args} />;
 
 Example.args = {
+  name: "email",
   children: (
     <div className="border border-gray-dim p-3 rounded-lg text-gray-dim">
       Children
