+++ src/components/forms/FormContainer/FormContainer.test.tsx
@@ -1,27 +1,42 @@
 import { FormContainer } from "@/components/forms";
 import { render, screen } from "@testing-library/react";
+import { useForm } from "react-hook-form";
 import { describe, expect, it } from "vitest";
+
+const FormContainerWithForm = () => {
+  const form = useForm();
+  return (
+    <FormContainer title="Test Title" form={form} onSubmit={() => {}}>
+      <div>Test Content</div>
+    </FormContainer>
+  );
+};
 
 describe("FormContainer", () => {
   it("renders children content", () => {
-    render(
-      <FormContainer>
-        <div>Test Content</div>
-      </FormContainer>,
-    );
+    render(<FormContainerWithForm />);
 
     expect(screen.getByText("Test Content")).toBeInTheDocument();
   });
 
+  it("renders title", () => {
+    render(<FormContainerWithForm />);
+
+    const title = screen.getByText("Test Title");
+
+    expect(title).toBeInTheDocument();
+    expect(title).toHaveClass("text-5xl");
+  });
+
   it("renders back button", async () => {
-    render(<FormContainer />);
+    render(<FormContainerWithForm />);
 
     const backButton = screen.getByLabelText("Save and exit");
     expect(backButton).toBeInTheDocument();
   });
 
   it("renders form navigation component", () => {
-    render(<FormContainer />);
+    render(<FormContainerWithForm />);
 
     // Check for navigation controls that FormNavigation renders
     expect(screen.getByLabelText("Previous question")).toBeInTheDocument();
