+++ src/components/forms/FormContainer/FormContainer.tsx
@@ -1,38 +1,82 @@
 import {
+  Banner,
   Button,
   Container,
+  Form,
   Tooltip,
   TooltipTrigger,
 } from "@/components/common";
 import { FormNavigation } from "@/components/forms";
+import { smartquotes } from "@/helpers/smartquotes";
 import { useRouter } from "@tanstack/react-router";
-import { ArrowLeft } from "lucide-react";
+import { ArrowLeft, Lock } from "lucide-react";
+import { Heading } from "react-aria-components";
+import { FormProvider, type UseFormReturn } from "react-hook-form";
 
 export interface FormContainerProps {
+  /** The title of the page. */
+  title: string;
+
+  /** An optional description to provide more context. */
+  description?: string;
+
   /** The contents of the page. */
   children?: React.ReactNode;
+
+  /** The form from react-hook-form's useForm hook. */
+  form: UseFormReturn<any>;
+
+  /** Submit handler for the form. */
+  onSubmit: React.FormEventHandler<HTMLFormElement>;
 }
 
-export function FormContainer({ children }: FormContainerProps) {
+export function FormContainer({
+  title,
+  description,
+  children,
+  form,
+  onSubmit,
+}: FormContainerProps) {
   const { history } = useRouter();
 
   return (
-    <main className="bg-gray-2 dark:bg-graydark-1">
-      <Container className="w-[720px] py-12">{children}</Container>
-      <TooltipTrigger>
-        <Button
-          className="fixed top-4 left-4 z-10"
-          onPress={() => {
-            history.go(-1);
-          }}
-          aria-label="Save and exit"
-          icon={ArrowLeft}
-          variant="icon"
-          size="large"
-        />
-        <Tooltip placement="right">Save and exit</Tooltip>
-      </TooltipTrigger>
-      <FormNavigation />
+    <main>
+      <FormProvider {...form}>
+        <Container className="w-[720px] py-16">
+          <Form onSubmit={onSubmit}>
+            <header className="flex flex-col gap-6 mb-8">
+              <Heading className="text-5xl font-medium text-pretty">
+                {title}
+              </Heading>
+              {description && (
+                <p className="text-sm text-gray-dim">
+                  {smartquotes(description)}
+                </p>
+              )}
+              <Banner variant="info" icon={Lock} size="large">
+                Namesake takes your privacy seriously. All responses are
+                end-to-end encrypted. That means no one—not even Namesake—can
+                see your answers.
+              </Banner>
+            </header>
+            {children}
+          </Form>
+        </Container>
+        <TooltipTrigger>
+          <Button
+            className="fixed top-4 left-4 z-10"
+            onPress={() => {
+              history.go(-1);
+            }}
+            aria-label="Save and exit"
+            icon={ArrowLeft}
+            variant="icon"
+            size="large"
+          />
+          <Tooltip placement="right">Save and exit</Tooltip>
+        </TooltipTrigger>
+        <FormNavigation />
+      </FormProvider>
     </main>
   );
 }
