+++ src/components/forms/FormSection/FormSection.stories.tsx
@@ -13,6 +13,7 @@ export default meta;
 export const Example = (args: FormSectionProps) => <FormSection {...args} />;
 
 Example.args = {
+  title: "Form Section",
   children: (
     <div className="border border-gray-dim p-3 rounded-lg text-gray-dim">
       Children
