+++ src/components/forms/FormSection/FormSection.tsx
@@ -67,7 +67,7 @@ export function FormSection({
       data-form-section
       data-testid="form-section"
       className={twMerge(
-        "flex flex-col gap-8 p-8 pb-9 justify-center outline outline-1 outline-graya-4 dark:outline-graydarka-4 shadow-sm rounded-2xl bg-white dark:bg-graydark-2",
+        "flex flex-col gap-8 p-8 pb-9 justify-center outline outline-1 outline-graya-3 dark:outline-graydarka-3 shadow-sm dark:shadow-md rounded-2xl bg-white dark:bg-graydark-2",
         className,
       )}
       // TODO: React-hook-form is supposed to prevent disabled fields from being submitted
