+++ src/components/forms/NameField/NameField.test.tsx
@@ -7,9 +7,9 @@ describe("NameField", () => {
   it("renders all name input fields", () => {
     renderWithFormProvider(<NameField type="newName" />);
 
-    const firstNameInput = screen.getByLabelText("First name");
+    const firstNameInput = screen.getByLabelText("First or given name");
     const middleNameInput = screen.getByLabelText("Middle name");
-    const lastNameInput = screen.getByLabelText("Last name");
+    const lastNameInput = screen.getByLabelText("Last or family name");
 
     expect(firstNameInput).toBeInTheDocument();
     expect(middleNameInput).toBeInTheDocument();
@@ -19,9 +19,9 @@ describe("NameField", () => {
   it("has correct autocomplete attributes", () => {
     renderWithFormProvider(<NameField type="newName" />);
 
-    const firstNameInput = screen.getByLabelText("First name");
+    const firstNameInput = screen.getByLabelText("First or given name");
     const middleNameInput = screen.getByLabelText("Middle name");
-    const lastNameInput = screen.getByLabelText("Last name");
+    const lastNameInput = screen.getByLabelText("Last or family name");
 
     expect(firstNameInput).toHaveAttribute("autocomplete", "given-name");
     expect(middleNameInput).toHaveAttribute("autocomplete", "additional-name");
@@ -31,11 +31,14 @@ describe("NameField", () => {
   it("allows entering text in all name fields", async () => {
     renderWithFormProvider(<NameField type="newName" />);
 
-    const firstNameInput: HTMLInputElement =
-      screen.getByLabelText("First name");
+    const firstNameInput: HTMLInputElement = screen.getByLabelText(
+      "First or given name",
+    );
     const middleNameInput: HTMLInputElement =
       screen.getByLabelText("Middle name");
-    const lastNameInput: HTMLInputElement = screen.getByLabelText("Last name");
+    const lastNameInput: HTMLInputElement = screen.getByLabelText(
+      "Last or family name",
+    );
 
     await userEvent.type(firstNameInput, "John");
     await userEvent.type(middleNameInput, "Michael");
