+++ src/components/forms/RadioGroupField/RadioGroupField.tsx
@@ -4,6 +4,8 @@ import {
   type RadioGroupProps,
   type RadioProps,
 } from "@/components/common";
+import type { UserFormDataField } from "@convex/constants";
+import { Controller, useFormContext } from "react-hook-form";
 
 interface RadioOption extends RadioProps {
   label: string;
@@ -12,7 +14,7 @@ interface RadioOption extends RadioProps {
 
 export interface RadioGroupFieldProps extends RadioGroupProps {
   children?: React.ReactNode;
-  name: string;
+  name: UserFormDataField;
   label: string;
   options: RadioOption[];
 }
@@ -22,17 +24,34 @@ export function RadioGroupField({
   label,
   options,
   children,
+  ...props
 }: RadioGroupFieldProps) {
+  const { control } = useFormContext();
+
   return (
     <div className="@container flex flex-col gap-4">
-      <RadioGroup label={label} name={name} size="large">
-        <span className="italic text-gray-dim text-sm">Select one:</span>
-        {options?.map(({ label, ...option }) => (
-          <Radio key={option.value} {...option} size="large" card>
-            {label}
-          </Radio>
-        ))}
-      </RadioGroup>
+      <Controller
+        control={control}
+        name={name}
+        render={({ field, fieldState: { invalid, error } }) => (
+          <RadioGroup
+            {...field}
+            label={label}
+            size="large"
+            isInvalid={invalid}
+            errorMessage={error?.message}
+            orientation="vertical"
+            {...props}
+          >
+            <span className="italic text-gray-dim text-sm">Select one:</span>
+            {options?.map(({ label, ...option }) => (
+              <Radio key={option.value} {...option} size="large" card>
+                {label}
+              </Radio>
+            ))}
+          </RadioGroup>
+        )}
+      />
       {children}
     </div>
   );
